// temp_deploy/sGED-frontend/src/app/core/models/busqueda.model.ts
export interface BusquedaAvanzadaRequest {
  numero?: string;
  fechaDesde?: string | Date;
  fechaHasta?: string | Date;
  estadoId?: number;
  tipoProcesoId?: number;
  juzgadoId?: number;
  actorPrincipal?: string;
  demandadoPrincipal?: string;
  referenciaSgt?: string;
}

export interface ExpedienteBusquedaResponse {
  id?: number | null;
  numero: string;
  juzgado: string;
  estado: string;
  tipoProceso: string;
  fechaInicio: string;
  fechaUltimoMovimiento?: string;
  fuente: 'SGED' | 'SGTV2' | 'SGTV1';
  actorPrincipal?: string;
  demandadoPrincipal?: string;
}